requirements.txt
prefect>=2.14.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.12.0
dbt-bigquery>=1.7.0
kagglehub>=0.3.0
